Tentative changes to Cyl.SurfacePlanarSurfaceInterfaceSingle
We want that the particle still stays on the cylinder in that situation,
to get the right unbinding behavior. However, to make this work we have
to get rid of the restrictions to the possible target structure types
in fire_single_reaction(), i.e. generalize the principle of going to
the bulk or parent structure by using the same structure functions
as in BD mode.